refactor(desktop): repoint FF context bridge to Redux, drop firebase remote-config provider
Make the live-common FeatureFlagsProvider bridge read feature-flag values from the Redux
slice (@shared/feature-flags) instead of LiveConfig, and remove the Firebase remote-config
provider. The bridge stays mounted (renamed FeatureFlagsContextBridge) so live-common's own
internal Context consumers keep resolving from the same source until Phase C migrates them;
its deletion is Phase D (LIVE-31228).

- FeatureFlagsContextBridge: getFeature reads selectFeature(resolved) reactively, isFeature
  uses FeatureIdSchema.safeParse; the now-dead imperative overrideFeature is simplified to a
  plain setOverride dispatch; the Wallet 4.0 graph-rework effect reads useWalletFeaturesConfig
  from @features/platform-feature-flags.
- firebase/remoteConfig.ts: add installLiveConfigProvider() (LiveConfig.setProvider via the
  RemoteConfig singleton + the relocated dev-only warnOnConfigMismatch).
- init.tsx: install LiveConfig + wire setAnalyticsFeatureFlagMethod from the slice after store
  creation, and gate boot on the feature-flags slice being ready (selectRemoteFlagsReady —
  dispatched by the middleware after the first fetch settles AND syncs into the slice, set
  unconditionally so a Firebase outage still releases the gate). Replaces the
  FirebaseRemoteConfigProvider null-gate and guarantees resolved flags are in the store before
  first render.
- App.tsx: swap the provider for the bridge; drop FirebaseRemoteConfigProvider.
- AppVersionBlocker: re-read config_ll_min_version via subscribeToRemoteFlags instead of the
  FirebaseRemoteConfig context.
- Delete FirebaseRemoteConfig.{tsx,test.tsx}, FirebaseFeatureFlags.{tsx,test.tsx},
  tests/featureFlags.ts; testSetup uses the bridge; repoint the guardrails exemption; add the
  bridge test.

LIVE-31639
