+++ apps/ledger-live-desktop/src/renderer/App.tsx
@@ -3,7 +3,6 @@ import { Provider } from "react-redux";
 import { useSelector } from "LLD/hooks/redux";
 import { Store } from "redux";
 import { HashRouter as Router } from "react-router";
-import { getFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { DeviceManagementKitProvider } from "@ledgerhq/live-dmk-desktop";
 import "./global.css";
 import "tippy.js/dist/tippy.css";
@@ -15,8 +14,7 @@ import StyleProvider from "~/renderer/styles/StyleProvider";
 import { UpdaterProvider } from "~/renderer/components/Updater/UpdaterContext";
 import ThrowBlock from "~/renderer/components/ThrowBlock";
 import LiveStyleSheetManager from "~/renderer/styles/LiveStyleSheetManager";
-import { FirebaseRemoteConfigProvider } from "~/renderer/components/FirebaseRemoteConfig";
-import { FirebaseFeatureFlagsProvider } from "~/renderer/components/FirebaseFeatureFlags";
+import { FeatureFlagsContextBridge } from "~/renderer/components/FeatureFlagsContextBridge";
 import { CountervaluesBridgedProvider } from "~/renderer/components/CountervaluesProvider";
 import DrawerProvider from "~/renderer/drawers/Provider";
 import Default from "./Default";
@@ -75,36 +73,34 @@ const InnerApp = ({ initialCountervalues }: { initialCountervalues: CounterValue
             }
           }}
         >
-          <FirebaseRemoteConfigProvider>
-            <FirebaseFeatureFlagsProvider getFeature={getFeature}>
-              <ConnectEnvsToSentry />
-              <ConnectEnvsToDatadog />
-              <UpdaterProvider>
-                <AppDataStorageProvider>
-                  <DeviceManagementKitProvider>
-                    <CountervaluesBridgedProvider initialState={initialCountervalues}>
-                      <ToastProvider>
-                        <ServiceStatusProviderWrapper>
-                          <Router>
-                            <PostOnboardingProviderWrapped>
-                              <PlatformAppProviderWrapper>
-                                <DrawerProvider>
-                                  <QueryClientProvider client={queryClient}>
-                                    <Default />
-                                    <ReactQueryDevtoolsProvider />
-                                  </QueryClientProvider>
-                                </DrawerProvider>
-                              </PlatformAppProviderWrapper>
-                            </PostOnboardingProviderWrapped>
-                          </Router>
-                        </ServiceStatusProviderWrapper>
-                      </ToastProvider>
-                    </CountervaluesBridgedProvider>
-                  </DeviceManagementKitProvider>
-                </AppDataStorageProvider>
-              </UpdaterProvider>
-            </FirebaseFeatureFlagsProvider>
-          </FirebaseRemoteConfigProvider>
+          <FeatureFlagsContextBridge>
+            <ConnectEnvsToSentry />
+            <ConnectEnvsToDatadog />
+            <UpdaterProvider>
+              <AppDataStorageProvider>
+                <DeviceManagementKitProvider>
+                  <CountervaluesBridgedProvider initialState={initialCountervalues}>
+                    <ToastProvider>
+                      <ServiceStatusProviderWrapper>
+                        <Router>
+                          <PostOnboardingProviderWrapped>
+                            <PlatformAppProviderWrapper>
+                              <DrawerProvider>
+                                <QueryClientProvider client={queryClient}>
+                                  <Default />
+                                  <ReactQueryDevtoolsProvider />
+                                </QueryClientProvider>
+                              </DrawerProvider>
+                            </PlatformAppProviderWrapper>
+                          </PostOnboardingProviderWrapped>
+                        </Router>
+                      </ServiceStatusProviderWrapper>
+                    </ToastProvider>
+                  </CountervaluesBridgedProvider>
+                </DeviceManagementKitProvider>
+              </AppDataStorageProvider>
+            </UpdaterProvider>
+          </FeatureFlagsContextBridge>
         </ThrowBlock>
       </ThemeProvider>
     </StyleProvider>
