+++ apps/ledger-live-desktop/tests/testSetup.tsx
@@ -23,7 +23,7 @@ import { MemoryRouter } from "react-router";
 import { config } from "react-transition-group";
 import ContextMenuWrapper from "~/renderer/components/ContextMenu/ContextMenuWrapper";
 import { useCountervaluesBridge } from "~/renderer/components/CountervaluesProvider";
-import { FirebaseFeatureFlagsProvider } from "~/renderer/components/FirebaseFeatureFlags";
+import { FeatureFlagsContextBridge } from "~/renderer/components/FeatureFlagsContextBridge";
 import type { ReduxStore } from "~/state-manager/configureStore";
 import createStore from "~/state-manager/configureStore";
 import DrawerProvider from "~/renderer/drawers/Provider";
@@ -34,7 +34,6 @@ import LiveStyleSheetManager from "~/renderer/styles/LiveStyleSheetManager";
 import StyleProvider from "~/renderer/styles/StyleProvider";
 import { RampCatalogProvider } from "@ledgerhq/live-common/platform/providers/RampCatalogProvider/index";
 import CustomLiveAppProvider from "./CustomLiveAppProvider";
-import { getFeature } from "./featureFlags";
 import { initialCountervaluesMock } from "./mocks/countervalues.mock";
 
 config.disabled = true;
@@ -175,7 +174,7 @@ function Providers({
   return (
     <QueryClientProvider client={queryClient}>
       <Provider store={store}>
-        <FirebaseFeatureFlagsProvider getFeature={getFeature}>
+        <FeatureFlagsContextBridge>
           {skipRouter ? (
             routerContent
           ) : (
@@ -183,7 +182,7 @@ function Providers({
               {routerContent}
             </MemoryRouter>
           )}
-        </FirebaseFeatureFlagsProvider>
+        </FeatureFlagsContextBridge>
       </Provider>
     </QueryClientProvider>
   );
